include: Fix compiler warnings for Windows build
Build logs are filled with the warning messages since the headers are included
for almost every source file.

windows\osd.h(777,26): warning C4090: 'function': different 'const' qualifiers
windows\osd.h(884,34): warning C4244: 'return': conversion from 'ULONGLONG' to 'long', possible loss of data
windows\osd.h(1039,19): warning C4057: 'function': 'int *' differs in indirection to slightly different base types from 'unsigned int *'
ofi_atom.h(362): warning C4057: 'function': 'volatile long *' differs in indirection to slightly different base types from 'int32_t *'
ofi_signal.h(115,19): warning C4244: '=': conversion from 'ssize_t' to 'int', possible loss of data
ofi_signal.h(147,18): warning C4244: '=': conversion from 'ssize_t' to 'int', possible loss of data
ofi_net.h(180,21): warning C4267: '+=': conversion from 'size_t' to 'unsigned int', possible loss of data
ofi_net.h(195,20): warning C4267: '+=': conversion from 'size_t' to 'unsigned int', possible loss of data
ofi_net.h(211,21): warning C4244: '+=': conversion from 'ssize_t' to 'unsigned int', possible loss of data
ofi_net.h(227,10): warning C4389: '==': signed/unsigned mismatch
ofi_net.h(231,21): warning C4244: '+=': conversion from 'ssize_t' to 'unsigned int', possible loss of data

+++ include/ofi_signal.h
@@ -105,7 +105,7 @@ static inline void fd_signal_set(struct fd_signal *signal)
 {
 	char c = 0;
 	bool cas; /* cas result */
-	int write_rc;
+	ssize_t write_rc;
 
 	cas = ofi_atomic_cas_bool_strong32(&signal->state,
 					   OFI_SIGNAL_UNSET,
@@ -136,7 +136,7 @@ static inline void fd_signal_reset(struct fd_signal *signal)
 	char c;
 	bool cas; /* cas result */
 	enum ofi_signal_state state;
-	int read_rc;
+	ssize_t read_rc;
 
 	do {
 		cas = ofi_atomic_cas_bool_weak32(&signal->state,
